pyproject.toml
[tool.poetry]
name = "testbot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.7.3"
discord-pretty-help = "^1.3.2"
keep-alive = "^0.0.1"
Flask = "^2.0.1"
logger = "^1.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
